Changes provided

+++ zero1one/src/main/java/com/gruppezwei/zero1one/repository/BerechtigungRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BerechtigungRepository extends JpaRepository<Berechtigung, Long> {
+public interface BerechtigungRepository extends JpaRepository<Berechtigung, Integer> {
 
 }
